Created Helper for Login methods - Sha/Nonce
