+++ LibTrack/Info.plist
@@ -26,6 +26,7 @@
 			<key>CFBundleURLSchemes</key>
 			<array>
 				<string>com.googleusercontent.apps.786408218821-np2tbohtkid1j3tbv6ctskrj4r5bodcg</string>
+				<string>com.googleusercontent.apps.786408218821-56c973mkbrg892jdmr7g6g4pg78gjpdf</string>
 			</array>
 		</dict>
 	</array>
